feat (TMS-18): mapped get_region_by_point function in repository

+++ spring-app/src/main/java/org/example/tmsserver/repository/RegionIndicatorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RegionIndicatorRepository extends JpaRepository<RegionIndicator, Long> {
+public interface RegionIndicatorRepository extends JpaRepository<RegionIndicator, Integer> {
 }
